Large number of changes
Multiple approaches to building longest useful phrases. None works
well and these are just to record the attempts before winnowing.

Also, this is a checkpoint before revamping the cde_analyzer top-level
script (the launcher) to minimizing loading (importing) of libraries
that won't be used, i.e., restrict loading to the action invoked.